fetchStats is now its own module
